# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A modern Go-based authentication system with magic link functionality using Clean Architecture principles.\nThis API provides secure authentication mechanisms including magic link authentication, JWT token management, and user profile management.",
        "title": "Trading Alchemist API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Trading Alchemist Team",
            "email": "team@tradingalchemist.dev"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/providers": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves a list of all active LLM providers supported by the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "List available providers",
                "responses": {
                    "200": {
                        "description": "Providers retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/trading-alchemist_internal_application_chat.ProviderResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/providers/available-models": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves all available models with their API key configuration status in a single optimized call",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "Get available models with API key status",
                "responses": {
                    "200": {
                        "description": "Available models retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/trading-alchemist_internal_application_chat.ProviderResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/providers/settings": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves all provider settings for the currently authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "List user's provider settings",
                "responses": {
                    "200": {
                        "description": "User provider settings retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/trading-alchemist_internal_application_chat.UserProviderSettingResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates or updates a provider setting (API key, base URL) for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Providers"
                ],
                "summary": "Create or update a provider setting",
                "parameters": [
                    {
                        "description": "Provider setting information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_application_chat.UpsertUserProviderSettingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provider setting saved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/trading-alchemist_internal_application_chat.UserProviderSettingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Provider not found",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "trading-alchemist_internal_application_chat.ModelResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "tags": {
                    "description": "Example: [\"LLM\", \"CHAT\", \"200K\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "trading-alchemist_internal_application_chat.ProviderResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/trading-alchemist_internal_application_chat.ModelResponse"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "trading-alchemist_internal_application_chat.UpsertUserProviderSettingRequest": {
            "type": "object",
            "required": [
                "api_key",
                "provider_id"
            ],
            "properties": {
                "api_base_override": {
                    "type": "string"
                },
                "api_key": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "provider_id": {
                    "type": "string"
                }
            }
        },
        "trading-alchemist_internal_application_chat.UserProviderSettingResponse": {
            "type": "object",
            "properties": {
                "api_base_override": {
                    "type": "string"
                },
                "api_key_set": {
                    "description": "Indicates if the API key is configured, without exposing the key.",
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "provider_display_name": {
                    "type": "string"
                },
                "provider_id": {
                    "type": "string"
                },
                "provider_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "trading-alchemist_internal_presentation_responses.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code\nexample: VALIDATION_ERROR",
                    "type": "string"
                },
                "details": {
                    "description": "Optional error details\nexample: The email field must be a valid email address",
                    "type": "string"
                },
                "message": {
                    "description": "Error message\nexample: Invalid email address",
                    "type": "string"
                }
            }
        },
        "trading-alchemist_internal_presentation_responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorDetail"
                        }
                    ]
                },
                "success": {
                    "description": "Success indicator (always false for errors)\nexample: false",
                    "type": "boolean"
                }
            }
        },
        "trading-alchemist_internal_presentation_responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data"
                },
                "message": {
                    "description": "Optional success message\nexample: Operation completed successfully",
                    "type": "string"
                },
                "success": {
                    "description": "Success indicator\nexample: true",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT token. Usage: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}